El_kahoot_terminado(creo)

+++ Assets/QAEndPanel.cs
@@ -17,7 +17,7 @@ public class QAEndPanel : MonoBehaviour
 
 
 
-    public void Initialize(int points)
+    public void Initialize(float points)
     {
 
         //panelImage.sprite = sprites[totalPoints];
